Revise Data::Record's MOP for efficiency
The primary concern is instances being roles, since we do an awkward
sort of pun with this. We can delegate to a class instead to skip steps
pertaining to doing and punning of the instance role during compilation.

For now, replace MetamodelX::RecordHOW with a new kind that will replace
it when complete. Deal with the things we need to deal with right off
the bat:

- Replicating most of MetamodelX::RecordHOW's state. This can now be
  immutable. We don't need %!parameters; parameterize the delegate
  beforehand.
- Naming. Store the anonymous ID separately from the name for a free
  index.
- REPRs. We take P6opaque in lieu of Uninstantiable because method
  dispatch will refer objects of *our* type to our delegate's methods,
  not the delegate itself, which expects to be able to create instances
  of whatever self may be.
- Delegating.
- Typechecking.