Let there be record instance metamethods
Despite instances being classes now, thos can't go there because they'll
be put in the class' HOW when we need them in the HOW delegating to it.
Metamodel::MetaMethodContainer mixes in metamethods to a HOW with a
role; use a similar approach here, and eliminate any Metamodel::Mixins
dependency along the way by keeping the delegate as a type parameter
instead of an attribute.

Any common behaviour pertaining to a record's fields can now be defined
as metamethods with a MetamodelX::RecordHOW role candidate. This allows
us to dedup instance logic and will allow us to eliminate the awkward
!field-op once Data::Record::Map gets its time to shine.