Rewrite Data::Record::Map for recent times
All but two tests pass because giving a field map directly to `{@ @}`
should be OK. Not doing so is a mitigation for bugs in the typing of its
candidates. Make it treat its arguments like { } would in general.

Simplifying this part of the codebase reveals a critical error in
Data::Record::Map. Its structural keyword does the exact opposite of
what it's supposed to do; namely, enforce a structure for a hash. It is
to be replaced, but not at the same time as all this.