Eliminate delegation of metamethods via the annotations library
This also eliminates the need for Metamodel::RecorderHOW in general
(though we keep it for now). This is a metarole that wants to subtype
Metamodel::ClassHOW like before, but cannot because no attribute values
can be forwarded to new_type (no, not even with dynamic defaults).
annotations gives us a sort of stash with a fixed location to place our
meta-"attributes" given the internals of its direct mode of operation.
Now Metamodel::RecorderHOW's methods can just be metamethods on the
various record instances.