Inline record types by failing typechecks with control exceptions
It's not that lists are more eager now that results in breakage, it's
that there are two ways to fulfill the definition of a record instance's
protocol in the case of a list, as with a map. We can treat the "does not
typecheck" property of consumptions and coercions as being a literal
absence of a type to check, leading to a type of a fixed width of values
to consider, or we can define it in terms of failed typechecks,
resulting in the slurpy-like behaviour we had before. Neither can be
defined in terms of one another, but there are ways to have both without
breaking a common type. For now, we should at least try to be somewhat
conservative before a v1.0.0 release.

Preserving the old behaviour is achievable with a more dynamic handling
of flunked typechecks during the lifting of a field. This resolves the
breakage in list tests, but should any of this be true, what's the deal
with the integration test breakage? There, we flunk the innermost
typecheck of the tuple directly, indicating that record types are
inlined with these changes.